requirements.txt
click==8.0.1
colorama==0.4.4
db==0.1.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
ipdb==0.13.9
itsdangerous==2.0.1
Jinja2==3.0.1
mysql-connector-python==8.0.26
mysqlclient==2.0.3
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
virtualenv==20.7.2
Werkzeug==2.0.1



